+++ packages/eslint-plugin-fiori-tools/src/constants.ts
@@ -5,3 +5,6 @@ export const COMMON_LABEL = 'com.sap.vocabularies.Common.v1.Label';
 export const UI_HIDDEN = 'com.sap.vocabularies.UI.v1.Hidden';
 export const UI_TEXT_ARRANGEMENT = 'com.sap.vocabularies.UI.v1.TextArrangement';
 export const UI_FIELD_GROUP = 'com.sap.vocabularies.UI.v1.FieldGroup';
+export const UI_FACETS = 'com.sap.vocabularies.UI.v1.Facets';
+export const UI_COLLECTION_FACET = 'com.sap.vocabularies.UI.v1.CollectionFacet';
+export const UI_REFERENCE_FACET = 'com.sap.vocabularies.UI.v1.ReferenceFacet';