+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/lrop_v4_cap/app/lrop_v4_cap/webapp/test/integration/pages/JourneyRunner.js
@@ -1,12 +1,12 @@
 sap.ui.define([
     "sap/fe/test/JourneyRunner",
-	"testNameSpace/lropv4cap/test/integration/pages/BooksList.gen",
-	"testNameSpace/lropv4cap/test/integration/pages/BooksObjectPage.gen"
+	"testnamespace/lropv4cap/test/integration/pages/BooksList.gen",
+	"testnamespace/lropv4cap/test/integration/pages/BooksObjectPage.gen"
 ], function (JourneyRunner, BooksListGenerated, BooksObjectPageGenerated) {
     'use strict';
 
     const runner = new JourneyRunner({
-        launchUrl: sap.ui.require.toUrl('testNameSpace/lropv4cap') + '/test/flp.html#app-preview',
+        launchUrl: sap.ui.require.toUrl('testnamespace/lropv4cap') + '/test/flp.html#app-preview',
         pages: {
 			onTheBooksListGenerated: BooksListGenerated,
 			onTheBooksObjectPageGenerated: BooksObjectPageGenerated