+++ packages/cf-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 1.1.5
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.13 → 2.2.0
+
+## 1.1.4
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/nodejs-utils 1.0.9 → 1.0.10
+
 ## 1.1.3
 
 ### Patch Changes